pull ECDSA utils from bitcointools
