+++ evm/vm/__init__.py
@@ -1,0 +1,3 @@
+from .evm import (  # noqa: F401
+    BaseEVM,
+)
