+++ evm/vm/flavors/__init__.py
@@ -1,0 +1,3 @@
+from .frontier import (  # noqa: F401
+    FrontierEVM,
+)
